Switch from Firecracker to Cloud Hypervisor guests
Cloud Hypervisor is a better-maintained derivative of Firecracker so this
is a relatively minor change. We add support for PCI virtio with virtio-mem
hotplug, virtio-fs and virtio-console instead of emulated 8250 serial.

Drop builds for arm64 guest kernels as these are hard to test on my
available systems, which either don't have GICv3 support or are already
virtual machines and don't offer nested virtualisation.